+++ deployment/src/main/java/io/quarkiverse/zeebe/devservices/ZeebeDevServicesStartResult.java
@@ -4,11 +4,13 @@ import java.io.Closeable;
 
 public class ZeebeDevServicesStartResult {
     public final String gateway;
+    public final String client;
     public final String hazelcast;
     final Closeable closeable;
 
-    public ZeebeDevServicesStartResult(String gateway, String hazelcast, Closeable closeable) {
+    public ZeebeDevServicesStartResult(String gateway, String client, String hazelcast, Closeable closeable) {
         this.gateway = gateway;
+        this.client = client;
         this.hazelcast = hazelcast;
         this.closeable = closeable;
     }
