+++ integration-tests/bpmn/src/main/java/io/quarkiverse/zeebe/it/bpmn/gateway/Input.java
@@ -1,0 +1,9 @@
+package io.quarkiverse.zeebe.it.bpmn.gateway;
+
+import io.quarkus.runtime.annotations.RegisterForReflection;
+
+@RegisterForReflection
+public class Input {
+
+    public boolean read;
+}
