+++ integration-tests/bpmn/src/main/java/io/quarkiverse/zeebe/it/bpmn/gateway/Parameter.java
@@ -1,0 +1,11 @@
+package io.quarkiverse.zeebe.it.bpmn.gateway;
+
+import io.quarkus.runtime.annotations.RegisterForReflection;
+
+@RegisterForReflection
+public class Parameter {
+
+    public String data;
+
+    public String info;
+}
