+++ integration-tests/bpmn/src/main/java/io/quarkiverse/zeebe/it/bpmn/sayhello/SayHelloBusinessService.java
@@ -1,0 +1,11 @@
+package io.quarkiverse.zeebe.it.bpmn.sayhello;
+
+import javax.enterprise.context.ApplicationScoped;
+
+@ApplicationScoped
+public class SayHelloBusinessService {
+
+    public String hello(String name) {
+        return "Hi, " + name;
+    }
+}
