+++ integration-tests/bpmn/src/main/java/io/quarkiverse/zeebe/it/bpmn/sayhello/SayHelloParameter.java
@@ -1,0 +1,12 @@
+package io.quarkiverse.zeebe.it.bpmn.sayhello;
+
+import io.quarkus.runtime.annotations.RegisterForReflection;
+
+@RegisterForReflection
+public class SayHelloParameter {
+
+    public String name;
+
+    public String message;
+
+}
