+++ integration-tests/bpmn/src/main/resources/bpmn/gateway.bpmn
