+++ integration-tests/bpmn/src/main/resources/bpmn/say_hello.bpmn
