+++ integration-tests/bpmn/src/test/java/io/quarkiverse/zeebe/it/bpmn/AbstractTest.java
@@ -1,0 +1,17 @@
+package io.quarkiverse.zeebe.it.bpmn;
+
+import io.quarkiverse.zeebe.test.ZeebeTestResource;
+import io.quarkus.test.common.QuarkusTestResource;
+import io.restassured.RestAssured;
+import io.restassured.filter.log.ResponseLoggingFilter;
+
+@QuarkusTestResource(ZeebeTestResource.class)
+public class AbstractTest {
+
+    //Configure the containers for the test
+    static {
+        RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
+        RestAssured.filters(new ResponseLoggingFilter());
+    }
+
+}
