+++ integration-tests/bpmn/src/test/java/io/quarkiverse/zeebe/it/bpmn/gateway/GatewayIT.java
@@ -1,0 +1,7 @@
+package io.quarkiverse.zeebe.it.bpmn.gateway;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class GatewayIT extends GatewayTest {
+}
