+++ integration-tests/bpmn/src/test/java/io/quarkiverse/zeebe/it/bpmn/health/HealthCheckIT.java
@@ -1,0 +1,7 @@
+package io.quarkiverse.zeebe.it.bpmn.health;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class HealthCheckIT extends HealthCheckTest {
+}
