+++ integration-tests/bpmn/src/test/java/io/quarkiverse/zeebe/it/bpmn/sayhello/SayHelloIT.java
@@ -1,0 +1,7 @@
+package io.quarkiverse.zeebe.it.bpmn.sayhello;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class SayHelloIT extends SayHelloTest {
+}
