+++ integration-tests/docker/src/main/resources/application.properties
@@ -1,0 +1,7 @@
+quarkus.zeebe.health.enabled=true
+
+%test.quarkus.zeebe.devservices.hazelcast.enabled=true
+
+quarkus.container-image.build=true
+quarkus.container-image.tag=latest
+quarkus.container-image.group=quarkiverse
