--- integration-tests/src/main/java/io/quarkiverse/zeebe/it/gateway/Input.java
@@ -1,9 +1,0 @@
-package io.quarkiverse.zeebe.it.gateway;
-
-import io.quarkus.runtime.annotations.RegisterForReflection;
-
-@RegisterForReflection
-public class Input {
-
-    public boolean read;
-}
