--- integration-tests/src/main/java/io/quarkiverse/zeebe/it/gateway/Parameter.java
@@ -1,11 +1,0 @@
-package io.quarkiverse.zeebe.it.gateway;
-
-import io.quarkus.runtime.annotations.RegisterForReflection;
-
-@RegisterForReflection
-public class Parameter {
-
-    public String data;
-
-    public String info;
-}
