--- integration-tests/src/main/java/io/quarkiverse/zeebe/it/sayhello/SayHelloBusinessService.java
@@ -1,11 +1,0 @@
-package io.quarkiverse.zeebe.it.sayhello;
-
-import javax.enterprise.context.ApplicationScoped;
-
-@ApplicationScoped
-public class SayHelloBusinessService {
-
-    public String hello(String name) {
-        return "Hi, " + name;
-    }
-}
