--- integration-tests/src/main/java/io/quarkiverse/zeebe/it/sayhello/SayHelloParameter.java
@@ -1,12 +1,0 @@
-package io.quarkiverse.zeebe.it.sayhello;
-
-import io.quarkus.runtime.annotations.RegisterForReflection;
-
-@RegisterForReflection
-public class SayHelloParameter {
-
-    public String name;
-
-    public String message;
-
-}
