--- integration-tests/src/test/java/io/quarkiverse/zeebe/it/gateway/GatewayIT.java
@@ -1,7 +1,0 @@
-package io.quarkiverse.zeebe.it.gateway;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-
-@QuarkusIntegrationTest
-public class GatewayIT extends GatewayTest {
-}
