--- integration-tests/src/test/java/io/quarkiverse/zeebe/it/health/HealthCheckIT.java
@@ -1,7 +1,0 @@
-package io.quarkiverse.zeebe.it.health;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-
-@QuarkusIntegrationTest
-public class HealthCheckIT extends HealthCheckTest {
-}
