--- integration-tests/src/test/java/io/quarkiverse/zeebe/it/sayhello/SayHelloIT.java
@@ -1,7 +1,0 @@
-package io.quarkiverse.zeebe.it.sayhello;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-
-@QuarkusIntegrationTest
-public class SayHelloIT extends SayHelloTest {
-}
